requirements.txt
tensorflow
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api
matplotlib
numpy
keras